refacto(Export): extract generation in module
